Add a backend with blog & now functionality
This is a pretty big commit, but the jist of it is that it adds a
backend written in Node. This allows blog content & now data to be
pulled directly from a db (currently mariadb). It's still a WIP but I'd
like to code on my laptop (and not be completely doomed if my PC dies
for whatever reason).